Change navbar to a reusable component (#5987)
In order to use navbar and footer theme component from
`@swmansion/t-rex-ui`, we need to rewrite current navbar to be more
simplified. To achieve that we moved landing background from `Navbar` to
`Hero` component and footer svg to `FooterBackground`.

<img width="1482" alt="Screenshot 2024-05-14 at 10 47 45"
src="https://github.com/software-mansion/react-native-reanimated/assets/59940332/3f1d4207-c02d-43df-a3e7-381ad4f62f88">
<img width="1482" alt="Screenshot 2024-05-14 at 10 47 57"
src="https://github.com/software-mansion/react-native-reanimated/assets/59940332/10675e93-8bd5-478a-bad4-ac7fd4ea8326">
