+++ docs/package.json
@@ -46,7 +46,8 @@
     "react-native-web": "^0.18.12",
     "source-map": "^0.7.4",
     "source-map-loader": "^4.0.1",
-    "usehooks-ts": "^2.9.1"
+    "usehooks-ts": "^2.9.1",
+    "theme-rex": "file:./swmansion-t-rex-ui-0.0.2.tgz"
   },
   "devDependencies": {
     "@docusaurus/module-type-aliases": "^2.4.3",
