+++ docs/src/components/FooterBackground/styles.module.css
@@ -1,0 +1,123 @@
+.sponsorsBackground {
+  position: absolute;
+  bottom: 0;
+
+  background: linear-gradient(
+      180deg,
+      var(--swm-white) 0%,
+      rgba(252, 252, 255, 0) 60.77%
+    ),
+    linear-gradient(
+      180deg,
+      var(--swm-yellow-light-20) 0%,
+      var(--swm-red-light-40) 50%,
+      var(--swm-purple-light-40) 100%
+    );
+  width: 100%;
+  height: 40vh;
+
+  z-index: -100;
+}
+
+[data-theme='dark'] .sponsorsBackground {
+  background: linear-gradient(
+    180deg,
+    var(--swm-navy) 0%,
+    var(--swm-purple-dark-100) 143%
+  );
+}
+
+.sponsorsBackgroundStars svg {
+  position: absolute;
+  bottom: 10%;
+
+  width: 100%;
+}
+
+.sponsorsBackgroundStars svg circle:nth-child(2n) {
+  animation: 2.9s starBlink infinite;
+}
+
+.sponsorsBackgroundStars svg circle:nth-child(2n + 1) {
+  animation: 2.4s starBlink infinite;
+}
+
+.sponsorsBackgroundStars svg circle:nth-child(3n) {
+  animation: 3.2s starBlink infinite;
+}
+
+.sponsorsBackgroundStars svg circle:nth-child(3n + 1) {
+  animation: 4.6s starBlink infinite;
+}
+
+.sponsorsBackgroundStars svg path:nth-child(2n) {
+  animation: 2.7s starBlink infinite;
+}
+
+.sponsorsBackgroundStars svg path:nth-child(2n + 1) {
+  animation: 2.1s starBlink infinite;
+}
+
+.sponsorsBackgroundStars svg path:nth-child(3n) {
+  animation: 3.6s starBlink infinite;
+}
+
+.sponsorsBackgroundStars svg path:nth-child(3n + 1) {
+  animation: 4.2s starBlink infinite;
+}
+
+@keyframes starBlink {
+  0% {
+    opacity: 1;
+  }
+
+  50% {
+    opacity: 0.15;
+  }
+
+  100% {
+    opacity: 1;
+  }
+}
+
+.starsContainer {
+  position: relative;
+  margin-top: 106px;
+}
+
+[class*='footerLanding'] {
+  margin-top: -106px;
+}
+
+@media (max-width: 996px) {
+  .starsContainer {
+    position: relative;
+    margin-top: 120px;
+  }
+
+  [class*='footerLanding'] {
+    margin-top: -120px;
+  }
+}
+
+@media (max-width: 700px) {
+  .starsContainer {
+    position: relative;
+    margin-top: 146px;
+  }
+
+  [class*='footerLanding'] {
+    margin-top: -146px;
+  }
+}
+
+@media (max-width: 375px) {
+  .starsContainer {
+    position: relative;
+    margin-top: 172px;
+  }
+
+  [class*='footerLanding'] {
+    margin-top: -172px;
+  }
+}
