+++ docs/src/pages/index.tsx
@@ -6,6 +6,7 @@ import HomepageStartScreen from '@site/src/components/Hero/StartScreen';
 import ReanimatedFeatures from '@site/src/components/ReanimatedFeatures';
 import Animations from '@site/src/components/Animations';
 import Testimonials from '@site/src/components/Testimonials';
+import FooterBackground from '../components/FooterBackground';
 import Sponsors from '@site/src/components/Sponsors';
 import HireUsSection from '../components/HireUsSection';
 
@@ -20,6 +21,7 @@ export default function Home(): JSX.Element {
         <Sponsors />
         <HireUsSection />
       </div>
+      <FooterBackground />
     </Layout>
   );
 }
