+++ docs/src/theme/Footer/Layout/index.js
@@ -2,7 +2,6 @@ import React from 'react';
 import clsx from 'clsx';
 import styles from './styles.module.css';
 
-import Stars from '@site/static/img/stars-footer.svg';
 import usePageType from '@site/src/hooks/usePageType';
 
 export default function FooterLayout({ style, links, logo, copyright }) {
@@ -13,13 +12,6 @@ export default function FooterLayout({ style, links, logo, copyright }) {
       className={clsx('footer', isLanding && styles.footerLanding, {
         'footer--dark': style === 'dark',
       })}>
-      {isLanding && (
-        <div className={styles.sponsorsBackground}>
-          <div className={styles.sponsorsBackgroundStars}>
-            <Stars />
-          </div>
-        </div>
-      )}
       <div className="container container-fluid">
         {links}
         {(logo || copyright) && (
