+++ docs/src/theme/Navbar/Layout/styles.module.css
@@ -14,55 +14,3 @@
   position: relative;
   background-color: transparent;
 }
-
-.heroBackground {
-  position: absolute;
-  top: 0;
-
-  background: linear-gradient(
-      357.23deg,
-      var(--swm-white) 2.25%,
-      rgba(252, 252, 255, 0) 53.95%
-    ),
-    linear-gradient(
-      180deg,
-      var(--swm-purple-light-60) 0%,
-      var(--swm-red-light-40) 40%,
-      var(--swm-yellow-light-40) 85%
-    );
-  width: 100%;
-  height: 100vh;
-  min-height: 70vw;
-  overflow: hidden;
-  z-index: -100;
-}
-
-[data-theme='dark'] .heroBackground {
-  background: linear-gradient(
-    180deg,
-    var(--swm-purple-dark-100) 0%,
-    var(--swm-navy) 85%
-  );
-}
-
-.heroBackground[data-announcement-bar='true'] {
-  top: var(--docusaurus-announcement-bar-height);
-}
-
-@media (max-width: 1124px) and (max-height: 768px) {
-  .heroBackground {
-    min-height: 768px;
-  }
-}
-
-@media (min-width: 680px) and (max-width: 996px) and (max-height: 996px) {
-  .heroBackground {
-    min-height: 1124px !important;
-  }
-}
-
-@media (min-width: 1600px) and (max-height: 996px) {
-  .heroBackground {
-    min-height: 70vw !important;
-  }
-}
